Renamed imported modules
Removed "spatial_ornstein_uhlenbeck" from extended modules.
